api/requirements.txt
flask==1.1.2
tweepy==3.9.0
python-dotenv==0.14.0
